Change to allow me to inject another IsolatingPlayerTeleporter in plasmid
